requirements.txt
streamlit>=1.28.0
openai>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
loguru>=0.7.0
pydantic-settings>=2.0.0

sentence-transformers>=2.2.2
FlagEmbedding>=1.1.0
langchain>=0.1.0
langchain-community>=0.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
scikit-learn>=1.3.0
numpy>=1.24.0
